Allow buffers whose modes don't have files to update direnv as well (#19)
I primarily intend to use this with shell-mode (with dirtrack.el set up to
properly keep my shell's default-directory up to date).

Note that since dirtrack tends to update default-directory *after* running a
command, the very next Emacs command you execute will not be in the context of
the new direnv. You can work around this with some advice on shell-cd:

  (defadvice shell-cd (after maybe-update activate)
    (direnv--maybe-update-environment))