smarter default behaviour of summary messages on environment change
this changes the default behaviour to show a summary of environment
changes in the minibuffer, but only if there actually are changes.

for interactive calls, a summary message is always shown, as before.
explicit user interaction always leads to direct user feedback.

on the other hand, when using direnv-mode, and for non-interactive
calls (e.g. from hooks), the logic to decide whether to show a summary
message is now smarter than before. as a result the
direnv-always-show-summary variable is now enabled by default.

in practice, the new behaviour will avoid automatic useless minibuffer
messages (‘direnv: no changes’) all the time when switching between
files in different directories, but within the same direnv controlled
project directory, which is what happened before this change.

closes #23.