emit warnings for direnv errors
a normal message will not be visible when switching from a functional
direnv directory to another broken one (e.g. unapproved .envrc), since
the unloading of the first directory will be shown as a summary
instead, causing the problem to go unnoticed. instead, use (warn ...)
which will pop up a window instead.