Improve handling of direnv output (#42)
Only parse JSON from direnv stdout. Output on stderr still goes into the *direnv* buffer, but will never be parsed.

When a problem occurs, emit a warning, since a normal message will not be visible. An example scenario is switching from a functional direnv directory to a broken one (e.g. unapproved .envrc). The unloading of the first directory will be shown as a summary message, and the message emitted about the problem will be overwritten and hence will go unnoticed. Instead, use (warn ...) which will pop up a window with more details instead.

Fixes #41.