use direnv executable path instead of repeatedly searching $PATH
even though the executable name is already detected and stored, it
wasn't actually used when invoking call-process, which would instead
search $PATH for the direnv executable prior to actually executing it.
this is very visible as a series of faccessat() syscalls in strace(1)
output.

when direnv-mode is enabled, switching buffers can cause direnv to be
executed, which can be quite often. not sure what the actual performance
benefits are here, but avoiding repeated work won't hurt. 🤞